new rules for tunes and chat

+++ index.js
@@ -20,8 +20,8 @@ mybot.on('message', function (message) {
   }
 
   // console.log(message.guild.name, ' - ', message.author.username, '(', message.channel.name, ')', ': ', message.content)
-  if (message.channel.name === 'your-tunes' && message.author.username !== 'Chefkoch' && message.attachments && _.size(message.attachments) < 1) {
-    const warnMessage = message.reply('if you want to post feedback, please use #your-tunes-chat')
+  if (message.channel.name === 'your-tunes' && message.author.username !== 'Chefkoch' && message.attachments && _.size(message.attachments) < 1 && !message.content.includes('http')) {
+    const warnMessage = message.reply('if you want to post feedback, please use #your-tunes-chat - message will be deleted soon!')
     // delete the messages after 20secs
     setTimeout(() => {
       warnMessage.then(msg => msg.delete())
@@ -29,8 +29,8 @@ mybot.on('message', function (message) {
     }, timeoutMessages)
   }
 
-  if (message.channel.name === 'your-tunes-chat' && message.author.username !== 'Chefkoch' && message.attachments && _.size(message.attachments) > 0) {
-    const warnMessage = message.reply('if you want to post a track, please use #your-tunes')
+  if (message.channel.name === 'your-tunes-chat' && message.author.username !== 'Chefkoch' && message.attachments && _.size(message.attachments) > 0 && message.content.includes('http')) {
+    const warnMessage = message.reply('if you want to post a track, please use #your-tunes - message will be deleted soon!')
     // delete the messages after 20secs
     setTimeout(() => {
       warnMessage.then(msg => msg.delete())
@@ -45,11 +45,9 @@ mybot.on('message', function (message) {
   if (message.channel.name === 'your-clips' && message.author.username !== 'Chefkoch' && message.attachments && _.size(message.attachments) > 0) {
     bwclips(message)
   }
-  if (message.channel.name === 'olisar' && message.author.username !== 'Chefkoch' && message.attachments && _.size(message.attachments) > 0) {
-    bwpresets(message)
-  }
-  if (message.channel.name === 'olisar') {
-    console.log(message)
+
+  if (message.channel.name === 'olisar' && message.attachments && _.size(message.attachments) < 1 && !message.content.includes('http')) {
+    message.reply('if you want to post a track, please use #your-tunes - message will be deleted soon!')
   }
 })
